[#11]fix: 판매자 확인 로직 accounts로 이동
